Refactored DirectX/OpenGL multi-peer conductor classes.
