+++ Samples/Server/Directx-SpinningCube/App.cpp
@@ -419,7 +419,7 @@ bool AppMain(BOOL stopping)
 		{
 			for each (auto pair in cond.Peers())
 			{
-				auto peer = pair.second;
+				auto peer = (DirectXPeerConductor*)pair.second.get();
 
 				// Retrieves remote peer data from map, create new if needed.
 				std::shared_ptr<RemotePeerData> peerData;
